Move _version to meta.
