+++ setup.py
@@ -63,7 +63,9 @@ def get_version_and_cmdclass(version_file):
     return globals_["__version__"], globals_["cmdclass"]()
 
 
-version, version_cmdclass = get_version_and_cmdclass("src/gluonts/_version.py")
+version, version_cmdclass = get_version_and_cmdclass(
+    "src/gluonts/meta/_version.py"
+)
 
 
 class TypeCheckCommand(distutils.cmd.Command):
