+++ src/gluonts/__init__.py
@@ -16,7 +16,7 @@
 
 from pkgutil import extend_path
 
-from ._version import __version__
+from .meta._version import __version__
 
 __all__ = ["__version__", "__path__"]
